@@ -20,7 +20,7 @@ allprojects {
     apply plugin: 'eclipse'
     apply plugin: 'java'
 
-    sourceCompatibility = 11
+    sourceCompatibility = 17
 }
 
 repositories {
