@@ -24,8 +24,8 @@ if (project.hasProperty("ci")) {
 
 group = 'com.example.catalog'
 version = '0.0.1-SNAPSHOT'
-sourceCompatibility = 11
-targetCompatibility = 11
+sourceCompatibility = 17
+targetCompatibility = 17
 
 configurations {
     compileOnly {
@@ -50,9 +50,7 @@ dependencies {
     runtimeOnly 'com.h2database:h2'
     runtimeOnly 'org.postgresql:postgresql'
     implementation 'org.liquibase:liquibase-core'
-    implementation 'org.springdoc:springdoc-openapi-ui:1.7.0'
-    implementation "org.zalando:problem-spring-web-starter:${problem_spring_web_version}"
-    implementation "org.apache.commons:commons-lang3:${commons_lang_version}"
+    implementation 'org.springdoc:springdoc-openapi-starter-webmvc-ui:2.1.0'
     implementation "commons-io:commons-io:${commons_io_version}"
 
     testImplementation 'org.springframework.boot:spring-boot-starter-test'
@@ -65,7 +63,6 @@ dependencies {
 dependencyManagement {
     imports {
         mavenBom "org.springframework.cloud:spring-cloud-dependencies:${spring_cloud_version}"
-        mavenBom "org.testcontainers:testcontainers-bom:${testcontainers_version}"
     }
 }
 
