@@ -24,8 +24,8 @@ if (project.hasProperty("ci")) {
 
 group = 'com.example.order'
 version = '0.0.1-SNAPSHOT'
-sourceCompatibility = 11
-targetCompatibility = 11
+sourceCompatibility = 17
+targetCompatibility = 17
 
 configurations {
     compileOnly {
@@ -46,14 +46,11 @@ dependencies {
     annotationProcessor 'org.springframework.boot:spring-boot-configuration-processor'
     compileOnly 'org.projectlombok:lombok'
     annotationProcessor 'org.projectlombok:lombok'
-    implementation "org.springframework.cloud:spring-cloud-starter-aws-messaging:2.2.6.RELEASE"
     implementation 'org.springframework.boot:spring-boot-starter-data-jpa'
     runtimeOnly 'com.h2database:h2'
     runtimeOnly 'org.postgresql:postgresql'
     implementation 'org.flywaydb:flyway-core'
-    implementation 'org.springdoc:springdoc-openapi-ui:1.7.0'
-    implementation "org.zalando:problem-spring-web-starter:${problem_spring_web_version}"
-    implementation "org.apache.commons:commons-lang3:3.13.0"
+    implementation 'org.springdoc:springdoc-openapi-starter-webmvc-ui:2.1.0'
     implementation "commons-io:commons-io:${commons_io_version}"
 
     testImplementation 'org.springframework.boot:spring-boot-starter-test'
@@ -61,13 +58,11 @@ dependencies {
     testImplementation "org.awaitility:awaitility"
     testImplementation "org.testcontainers:junit-jupiter"
     testImplementation "org.testcontainers:postgresql"
-    testImplementation "org.testcontainers:localstack"
 }
 
 dependencyManagement {
     imports {
         mavenBom "org.springframework.cloud:spring-cloud-dependencies:${spring_cloud_version}"
-        mavenBom "org.testcontainers:testcontainers-bom:${testcontainers_version}"
     }
 }
 
