@@ -6,7 +6,7 @@ plugins {
 
 group = 'com.example.serviceregistry'
 version = '0.0.1-SNAPSHOT'
-sourceCompatibility = '11'
+sourceCompatibility = '17'
 
 repositories {
     mavenCentral()
